requirements.txt
instagrapi==2.0.0
pillow==10.2.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
feedparser==6.0.11
beautifulsoup4==4.12.3
moviepy==1.0.3
numpy==1.26.3
